API References

LocalBinaryPatterns.lbp_original — Method

lbp_original(X)
lbp_original!(out, X)

Compute the local binary pattern, the original version [1], of gray image using 3x3 neighborhood matrix.

3x3 block     center-thresholded     weights         multiplied by weights      sum
6  7  9         1  1  1              1  8  32          1   8  32
5  6  3  ==>    0  x  0     ==>      2  x  64  ==>     0   x  0            ==>  169
2  1  7         1  0  1              4  16 128         0   0  128

julia> X = [6 7 9; 5 6 3; 2 1 7]
3×3 Matrix{Int64}:
 6  7  9
 5  6  3
 2  1  7

julia> lbp_original(X)
3×3 Matrix{UInt8}:
 0xc0  0x40  0x00
 0x68  0xa9  0x1b
 0x28  0x6b  0x00

References

[1] T. Ojala, M. Pietikäinen, and D. Harwood, “A comparative study of texture measures with classification based on featured distributions,” Pattern Recognition, vol. 29, no. 1, pp. 51–59, Jan. 1996, doi: 10.1016/0031-3203(95)00067-4.

[2] T. Ojala, M. Pietikäinen, and T. Mäenpää, “A Generalized Local Binary Pattern Operator for Multiresolution Gray Scale and Rotation Invariant Texture Classification,” in Advances in Pattern Recognition — ICAPR 2001, vol. 2013, S. Singh, N. Murshed, and W. Kropatsch, Eds. Berlin, Heidelberg: Springer Berlin Heidelberg, 2001, pp. 399–408. doi: 10.1007/3-540-44732-6_41.